feat (front): finalizando o css dos modais de produtos
