--- front/src/renderer/src/views/pages/Menu/components/Products/NewProductModal/index.tsx
@@ -1,129 +1,0 @@
-import Modal from '@renderer/views/components/Modal';
-
-import { capitalizeFirstLetter } from '@renderer/app/utils/capitalizeFirstLetter';
-import Button from '@renderer/views/components/Button';
-import { Input } from '@renderer/views/components/Input';
-import Loader from '@renderer/views/components/Loader';
-import { CgCloseO } from 'react-icons/cg';
-import Ingredients from '../../Ingredients';
-import NewIngredientModal from '../../Ingredients/NewIngredientModal';
-import UploadImage from '../UploadImage';
-import {
-  CategoryItem,
-  CategoryList,
-  CategorySection,
-  Container,
-  Content,
-  Footer,
-  ImageSection,
-  InputsContainer,
-} from './styles';
-import useNewProductModal from './useNewProductModal';
-
-interface NewProductModalProps {
-  open: boolean
-  onClose?(): void
-}
-
-export default function NewProductModal({ open, onClose }: NewProductModalProps) {
-  const {
-    errors,
-    width,
-    categories,
-    isLoadingCategories,
-    selectedCategory,
-    previewImageUrl,
-    openNewIngredientModal,
-    register,
-    handleSelectedCategory,
-    handleSelectedIngredients,
-    handleSubmit,
-    handleUploadImage,
-    handleCloseNewIngredientModal,
-    handleOpenNewIngredientModal,
-  } = useNewProductModal();
-
-  if (openNewIngredientModal) {
-    return <NewIngredientModal open onClose={handleCloseNewIngredientModal} />;
-  }
-
-  return (
-    <Modal
-      title="Novo produto"
-      open={open}
-      onClose={onClose}
-      $maxWidth={width < 1024 ? '500px' : '928px'}
-    >
-      <Container>
-        <Content>
-          <ImageSection>
-            <h3>Imagem</h3>
-
-            <UploadImage onUpload={handleUploadImage} previewImageUrl={previewImageUrl} />
-          </ImageSection>
-
-          <InputsContainer>
-            <Input
-              type="text"
-              placeholder="Nome do produto"
-              {...register('name')}
-              $error={errors.name?.message}
-            />
-
-            <Input
-              type="text"
-              placeholder="Descrição"
-              maxLength={110}
-              {...register('description')}
-              $error={errors.description?.message}
-            />
-
-            <p>Máximo 110 caracteres</p>
-          </InputsContainer>
-
-          <CategorySection>
-            <header>
-              <p className="title">Categoria</p>
-
-              {errors.category && (
-                <span className="error">
-                  <CgCloseO color="#F63131" />
-                  {errors.category.message}
-                </span>
-              )}
-            </header>
-
-            <CategoryList>
-              {isLoadingCategories && (
-                <div className='categories-loader'>
-                  <Loader size={14} $isLoading/>
-                </div>
-              )}
-
-              {!isLoadingCategories && (
-                categories.map((categorie, key) => (
-                  <CategoryItem
-                    key={key}
-                    $selected={categorie.id === selectedCategory?.id}
-                    onClick={() => handleSelectedCategory(categorie)}
-                    tabIndex={0} // Permite navegação por teclado
-                    onKeyDown={(e) => e.key === 'Enter' && handleSelectedCategory(categorie)}
-                  >
-                    <span>{categorie.icon}</span>
-                    <span>{capitalizeFirstLetter(categorie.name)}</span>
-                  </CategoryItem>
-                ))
-              )}
-            </CategoryList>
-          </CategorySection>
-
-          <Ingredients openNewIngredientModal={handleOpenNewIngredientModal} onSelected={handleSelectedIngredients}/>
-        </Content>
-
-        <Footer>
-          <Button onClick={handleSubmit}>Adicionar produtos</Button>
-        </Footer>
-      </Container>
-    </Modal>
-  );
-}
