+++ front/src/renderer/src/views/pages/Menu/components/Products/index.tsx
@@ -1,56 +1,39 @@
 import { products } from '@renderer/app/mocks/products';
-import { Table } from '../Table';
 
-import { Product } from '@renderer/app/entities/Product';
 import { capitalizeFirstLetter } from '@renderer/app/utils/capitalizeFirstLetter';
 import { formatCurrency } from '@renderer/app/utils/formatCurrency';
 import { DeleteIcon } from '@renderer/assets/Icons/DeleteIcon';
 import { Pencil } from '@renderer/assets/Icons/Pencil';
-import { useCallback, useState } from 'react';
-import DeleteProductModal from './DeleteProductModal';
-import EditProductModal from './EditProductModal';
-import NewProductModal from './NewProductModal';
+
+import { Table } from '../Table';
+import DeleteProductModal from './modals/DeleteProductModal';
+import EditProductModal from './modals/EditProductModal';
+import NewProductModal from './modals/NewProductModal';
+
+import useProducts from './useProducts';
+
 import { ActionButton, CategoryContainer, Container, Header, ProductImage, Title } from './styles';
 
-export default function Produtos() {
-  const [openEditModal, setOpenEditModal] = useState(true);
-  const [openDeleteModal, setOpenDeleteModal] = useState(false);
-  const [openNewProductModal, setOpenNewProductModal] = useState(true);
-
-  const [productBeingDeleted, setProductBeingDeleted] = useState<Product | null>(null);
-  const [productBeingEdited, setproductBeingEdited] = useState<Product | null>(null);
-
-  const handleOpenEditModal = useCallback((product: Product) => {
-    setproductBeingEdited(product);
-    setOpenEditModal(true);
-  }, []);
-
-  const handleCloseEditModal = useCallback(() => {
-    setproductBeingEdited(null);
-    setOpenEditModal(false);
-  }, []);
-
-  const handleOpenDeleteModal = useCallback((product: Product) => {
-    setProductBeingDeleted(product);
-    setOpenDeleteModal(true);
-  }, []);
-
-  const handleCloseDeleteModal = useCallback(() => {
-    setproductBeingEdited(null);
-    setOpenDeleteModal(false);
-  }, []);
-
-  const handleOpenNewProductModal = useCallback(() => {
-    setOpenNewProductModal(true);
-  }, []);
-
-  const handleCloseNewProductModal = useCallback(() => {
-    setOpenNewProductModal(false);
-  }, []);
+export default function Products() {
+  const {
+    productBeingDeleted,
+    productBeingEdited,
+    openEditModal,
+    openDeleteModal,
+    openNewProductModal,
+    handleOpenEditModal,
+    handleOpenDeleteModal,
+    handleOpenNewProductModal,
+    handleCloseEditModal,
+    handleCloseDeleteModal,
+    handleCloseNewProductModal,
+  } = useProducts();
 
   return (
     <>
-      <NewProductModal open={openNewProductModal} onClose={handleCloseNewProductModal} />
+      {openNewProductModal && (
+        <NewProductModal open={openNewProductModal} onClose={handleCloseNewProductModal} />
+      )}
 
       {productBeingEdited && (
         <EditProductModal
