+++ front/src/renderer/src/views/pages/Menu/components/Products/modals/DeleteProductModal/styles.ts
