+++ front/src/renderer/src/views/pages/Menu/components/Products/modals/NewProductModal/useNewProductModal.ts
@@ -1,12 +1,10 @@
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
 import { Ingredient } from '@renderer/app/entities/Ingredient';
 import { ProductCategory } from '@renderer/app/entities/ProductCategory';
-import { useIngredients } from '@renderer/app/hooks/queries/useIngredients';
-import { useProductCategories } from '@renderer/app/hooks/queries/useProductCategories';
 import { useWindowWidth } from '@renderer/app/hooks/useWindowWidth';
 
 const schema = z.object({
@@ -39,9 +37,6 @@ export default function useNewProductModal() {
   const [openNewIngredientModal, setOpenNewIngredientModal] = useState(false);
 
   const width = useWindowWidth();
-
-  const { ingredients, isLoading: isLoadingIngredients } = useIngredients();
-  const { categories, isLoading: isLoadingCategories } = useProductCategories();
 
   const {
     register,
@@ -97,16 +92,9 @@ export default function useNewProductModal() {
     console.log(data);
   });
 
-  useEffect(() => {
-  }, [selectedCategory]);
-
   return {
     errors,
     width,
-    ingredients,
-    isLoadingIngredients,
-    categories,
-    isLoadingCategories,
     selectedCategory,
     selectedIngredients,
     previewImageUrl,
